% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods.R
\name{read_ods_cpp}
\alias{read_ods_cpp}
\title{Read the data from an ODS file}
\usage{
read_ods_cpp(
  path,
  sheet = 1,
  startcol = 1,
  stopcol = -1,
  startrow = 1,
  stoprow = -1
)
}
\description{
Read the data from an ODS file
}
