% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ods_sheets.R
\name{list_ods_sheets}
\alias{list_ods_sheets}
\title{List all sheets in an ODS File}
\usage{
list_ods_sheets(path)
}
\arguments{
\item{path}{Path to the ods file}
}
\value{
A character vector of sheet names.
}
\description{
List all shgeets in an ods file.
}
