% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods.R
\name{read_ods_cpp}
\alias{read_ods_cpp}
\title{Read the data from an ODS file}
\usage{
read_ods_cpp(
  path,
  sheet = 1,
  col_names = TRUE,
  col_types = NULL,
  na = "",
  skip = 0,
  formula_as_formula = FALSE,
  range = NULL,
  row_names = FALSE,
  strings_as_factors = FALSE,
  check_names = FALSE,
  verbose = FALSE
)
}
\description{
Read the data from an ODS file
}
